requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
requests==2.32.0
psutil==6.1.0
gputil==1.4.0
pydantic==2.6.4
python-multipart==0.0.9

# Added for caching, task queue, observability, security, and compression
redis==5.0.0
aioredis==2.0.1
celery==5.3.4
slowapi==0.1.9
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-jaeger==1.21.0
compression==0.1.0

# testing utilities
pytest==7.3.2
httpx==0.27.0

# Authentication & security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Vector database & semantic search
sentence-transformers==2.2.2
pinecone-client==3.0.0
faiss-cpu==1.7.4

# RAG & LLM orchestration
langchain==0.1.0
langchain-community==0.0.13
PyPDF2==3.0.1

# Analytics & NLP
spacy==3.7.2
textblob==0.17.1
scikit-learn==1.4.0

# WebSocket & real-time
websockets==12.0
python-socketio==5.10.0
python-engineio==4.8.0

# API & schema
content-range==1.1.0

# CLI
click==8.1.7
click-rich==1.1.0
rich==13.7.0
tabulate==0.9.0